requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2
django-cors-headers==3.6.0
django-filter==23.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-yasg==1.21.5
dynaconf==3.2.0
et-xmlfile==1.1.0
frozenlist==1.4.0
hvac==1.1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyhcl==0.4.4
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
validators==0.20.0
yarl==1.9.2
